Attempt to fix false beacon data: deleteOldest
Made the data worse, but thought id save it here for reference.

Explanation: invalidating the rssi value from the oldest signal eliminated false data from beacons no longer in range but also reset beacons in range to -100 causing fluctuations in the closestBeacon because it had to wait for a new signal to replace the -100 even though it was still the closest and in range.